[AIG] Refactor longest path analysis to enable caching and improve modularity
- Add LongestPathAnalysisWithTrace wrapper class to enable analysis manager
  caching by default enabling trace functionality
- Move PrintLongestPathAnalysisPass implementation to separate file
  (PrintLongestPathAnalysis.cpp) for better separation of concerns
- Update pass to use cached analysis results via getAnalysis<> instead of
  creating new analysis instances
