+++ lib/Dialect/AIG/Analysis/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_circt_dialect_library(CIRCTAIGAnalysis
   LongestPathAnalysis.cpp
+  PrintLongestPathAnalysis.cpp
 
   DEPENDS
   CIRCTAIGPassesIncGen
